requirements.txt
asgiref==3.5.2
Django==4.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
pip==23.3.1
setuptools==68.2.2
sqlparse==0.4.4
tzdata==2023.3
wheel==0.41.2
